user info from db by cookie

+++ server/ts/routes/toDoTask/eraseTask/index.ts
@@ -6,15 +6,23 @@ import Subtask from '../../../models/SubTask';
 const route = Router();
 
 route.post('/', async (req: Request, res: Response) => {
+    const userCookie = req.cookies.tasksListUbi;
+
     if(req.headers['x-http-method-override']  === 'DELETE') {
 
         try {
             const { id } = req.body;
 
-            await ToDo.deleteOne({ _id: id });
-            await Subtask.deleteMany({ parentId: id });
+            const originalUser = await ToDo.findOne({  _id: id, userCookie });
 
-            res.status(202).json({ message: `Successfully deleted` });
+            if(originalUser === null) {
+                throw Error('You are not the original creator of this task so you cannot delete it');
+            } else {
+                await ToDo.deleteOne({ _id: id });
+                await Subtask.deleteMany({ parentId: id });
+
+                res.status(202).json({ message: `Successfully deleted` });
+            }
         } catch(err) {
             console.error(err);
             res.status(500).json({ errorMessage: err.message });
